+++ src/Fields/Checkbox.php
@@ -17,6 +17,7 @@ use WordPlate\Acf\Fields\Attributes\Choices;
 use WordPlate\Acf\Fields\Attributes\ConditionalLogic;
 use WordPlate\Acf\Fields\Attributes\DefaultValue;
 use WordPlate\Acf\Fields\Attributes\DirectionLayout;
+use WordPlate\Acf\Fields\Attributes\Disabled;
 use WordPlate\Acf\Fields\Attributes\Instructions;
 use WordPlate\Acf\Fields\Attributes\Required;
 use WordPlate\Acf\Fields\Attributes\ReturnFormat;
@@ -26,6 +27,7 @@ class Checkbox extends Field
 {
     use Choices;
     use DefaultValue;
+    use Disabled;
     use DirectionLayout;
     use ConditionalLogic;
     use Instructions;