+++ src/Fields/DatePicker.php
@@ -15,6 +15,7 @@ namespace WordPlate\Acf\Fields;
 
 use WordPlate\Acf\Fields\Attributes\ConditionalLogic;
 use WordPlate\Acf\Fields\Attributes\DateTimeFormat;
+use WordPlate\Acf\Fields\Attributes\Disabled;
 use WordPlate\Acf\Fields\Attributes\Instructions;
 use WordPlate\Acf\Fields\Attributes\Required;
 use WordPlate\Acf\Fields\Attributes\WeekDay;
@@ -24,6 +25,7 @@ class DatePicker extends Field
 {
     use ConditionalLogic;
     use DateTimeFormat;
+    use Disabled;
     use Instructions;
     use Required;
     use WeekDay;