+++ src/Fields/Textarea.php
@@ -16,6 +16,7 @@ namespace WordPlate\Acf\Fields;
 use WordPlate\Acf\Fields\Attributes\CharacterLimit;
 use WordPlate\Acf\Fields\Attributes\ConditionalLogic;
 use WordPlate\Acf\Fields\Attributes\DefaultValue;
+use WordPlate\Acf\Fields\Attributes\Disabled;
 use WordPlate\Acf\Fields\Attributes\Instructions;
 use WordPlate\Acf\Fields\Attributes\NewLines;
 use WordPlate\Acf\Fields\Attributes\Placeholder;
@@ -27,6 +28,7 @@ class Textarea extends Field
     use CharacterLimit;
     use ConditionalLogic;
     use DefaultValue;
+    use Disabled;
     use Instructions;
     use NewLines;
     use Placeholder;